+++ LookupAnything/Framework/Lookups/Items/ItemSubject.cs
@@ -57,6 +57,9 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         /// <summary>Whether to highlight item gift tastes which haven't been revealed in the NPC profile.</summary>
         private readonly bool HighlightUnrevealedGiftTastes;
 
+        /// <summary>Provides subject entries.</summary>
+        private readonly ISubjectRegistry Codex;
+
         /// <summary>Get a lookup subject for a crop.</summary>
         private readonly Func<Crop, ObjectContext, HoeDirt, ISubject> GetCropSubject;
 
@@ -65,6 +68,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
+        /// <param name="codex">Provides subject entries</param>
         /// <param name="gameHelper">Provides utility methods for interacting with the game code.</param>
         /// <param name="progressionMode">Whether to only show content once the player discovers it.</param>
         /// <param name="highlightUnrevealedGiftTastes">Whether to highlight item gift tastes which haven't been revealed in the NPC profile.</param>
@@ -74,9 +78,10 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         /// <param name="getCropSubject">Get a lookup subject for a crop.</param>
         /// <param name="fromCrop">The crop associated with the item (if applicable).</param>
         /// <param name="fromDirt">The dirt containing the crop (if applicable).</param>
-        public ItemSubject(GameHelper gameHelper, bool progressionMode, bool highlightUnrevealedGiftTastes, Item item, ObjectContext context, bool knownQuality, Func<Crop, ObjectContext, HoeDirt, ISubject> getCropSubject, Crop fromCrop = null, HoeDirt fromDirt = null)
+        public ItemSubject(ISubjectRegistry codex, GameHelper gameHelper, bool progressionMode, bool highlightUnrevealedGiftTastes, Item item, ObjectContext context, bool knownQuality, Func<Crop, ObjectContext, HoeDirt, ISubject> getCropSubject, Crop fromCrop = null, HoeDirt fromDirt = null)
             : base(gameHelper)
         {
+            this.Codex = codex;
             this.ProgressionMode = progressionMode;
             this.HighlightUnrevealedGiftTastes = highlightUnrevealedGiftTastes;
             this.Target = item;
@@ -534,7 +539,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                         .ToArray();
 
                     // display fields
-                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj);
+                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, this.Codex);
                     if (minutesLeft <= 0 || !schedule.Any())
                         yield return new GenericField(I18n.Item_CaskSchedule(), I18n.Item_CaskSchedule_Now(quality: I18n.For(curQuality)));
                     else
@@ -556,7 +561,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                 if (heldObj == null)
                 {
                     if (pot.bait.Value != null)
-                        yield return new ItemIconField(this.GameHelper, I18n.Item_CrabpotBait(), pot.bait.Value);
+                        yield return new ItemIconField(this.GameHelper, I18n.Item_CrabpotBait(), pot.bait.Value, this.Codex);
                     else if (Game1.player.professions.Contains(11)) // no bait needed if luremaster
                         yield return new GenericField(I18n.Item_CrabpotBait(), I18n.Item_CrabpotBaitNotNeeded());
                     else
@@ -567,7 +572,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                 if (heldObj != null)
                 {
                     string summary = I18n.Item_Contents_Ready(name: heldObj.DisplayName);
-                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, summary);
+                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, this.Codex, summary);
                 }
             }
 
@@ -578,7 +583,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                 if (heldObj != null)
                 {
                     string summary = I18n.Item_Contents_Placed(name: heldObj.DisplayName);
-                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, summary);
+                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, this.Codex, summary);
                 }
             }
 
@@ -599,7 +604,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                     string summary = minutesLeft <= 0
                     ? I18n.Item_Contents_Ready(name: heldObj.DisplayName)
                     : I18n.Item_Contents_Partial(name: heldObj.DisplayName, time: this.Stringify(TimeSpan.FromMinutes(minutesLeft)));
-                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, summary);
+                    yield return new ItemIconField(this.GameHelper, I18n.Item_Contents(), heldObj, this.Codex, summary);
                 }
             }
         }