+++ LookupAnything/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Added support for looking up NPCs at festivals.
 * Added target redirection. This fixes an issue where looking up Abigail in the mines wouldn't show her real data. (See [`EnableTargetRedirection` in config](README.md#configure) to disable this fix.)
+* Item icon fields now link to their item's lookup.
 * Fixed furniture placed outside cabins/farmhouses/sheds not counted for 'owned' stats.
 * Fixed mahogany tree lookups showing wrong percentage growth chance.